refactor: Refactor frontend upload modal
separating the upload logics from the api routing
